% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cordex_utils.R
\name{get_cordex_map}
\alias{get_cordex_map}
\title{Create the main interactive map of the Nile Basin with the station locations}
\usage{
get_cordex_map()
}
\value{
Leaflet map widget
}
\description{
Create the main interactive map of the Nile Basin with the station locations
}
