% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cordex_utils.R
\name{get_cordex_monthly}
\alias{get_cordex_monthly}
\title{Get datasets that will be used in the plots}
\usage{
get_cordex_monthly(variable, period)
}
\description{
Get datasets that will be used in the plots
}
