% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cordex_utils.R
\name{get_cordex_monthly_plot}
\alias{get_cordex_monthly_plot}
\title{Generate the seasonality and annual plots}
\usage{
get_cordex_monthly_plot(vls_monthly, vls_annual)
}
\description{
Generate the seasonality and annual plots
}
