% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cordex_utils.R
\name{get_cordex_monthly_values}
\alias{get_cordex_monthly_values}
\title{Extract variables to be used in the plots}
\usage{
get_cordex_monthly_values(variable, period, x, y)
}
\description{
Extract variables to be used in the plots
}
