% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_evap_utils.R
\name{get_evap_map}
\alias{get_evap_map}
\title{Create the evaporation interactive map of the OWASA portal}
\usage{
get_evap_map()
}
\value{
Leaflet map widget
}
\description{
Create the evaporation interactive map of the OWASA portal
}
