% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_parters_utils.R
\name{get_partner_map}
\alias{get_partner_map}
\title{define the markers of stkeholders to the map}
\usage{
get_partner_map()
}
\value{
Leaflet map with stakeholders locations
}
\description{
define the markers of stkeholders to the map
}
