% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prec_utils.R
\name{get_prec_map}
\alias{get_prec_map}
\title{Create the precipitation interactive map of the OWASA portal}
\usage{
get_prec_map()
}
\value{
Leaflet map widget
}
\description{
Create the precipitation interactive map of the OWASA portal
}
