% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sm_utils.R
\name{get_sm_map}
\alias{get_sm_map}
\title{Create the soil moisture interactive map of the OWASA portal}
\usage{
get_sm_map()
}
\value{
Leaflet map widget
}
\description{
Create the soil moisture interactive map of the OWASA portal
}
